Pull in the code from my recent Ludum Dare attempt
The idea is to combine the ideas from both codebases, fix the issues
in the LD42 code for things like the scrolling screen and collision,
and then add some flags that the example game states can set to
trigger the proper functionality from the boilerplate.

It takes a few code changes to swap between the two examples, but that
will improve in future commits.
- Change the game state imported in main.py
- Change the game screen imported in boilerplate/game.py
- Change the run function in boilerplate/game.py